move SamplesScannedConfig to PlannerParams
