+++ coordinator/src/main/scala/filodb/coordinator/ProtoConverters.scala
@@ -16,7 +16,6 @@ import filodb.coordinator.flight.FlightPlanDispatcher
 import filodb.core.downsample.{CounterDownsamplePeriodMarker, TimeDownsamplePeriodMarker}
 import filodb.core.memstore.PartLookupResult
 import filodb.core.metadata.{ComputedColumn, DataColumn}
-import filodb.core.metrics.FilodbMetrics
 import filodb.core.query._
 import filodb.core.store.{AllChunkScan, InMemoryChunkScan, TimeRangeChunkScan, WriteBufferChunkScan}
 import filodb.grpc.GrpcMultiPartitionQueryService
@@ -34,9 +33,6 @@ import filodb.query.exec._
 object ProtoConverters extends StrictLogging {
 
   import filodb.query.ProtoConverters._
-
-  // Increments when a fully-qualified class name is sent as a proto field but not found via reflection.
-  val unexpectedProtoClassNameCounter = FilodbMetrics.counter("unexpected-proto-class-name")
 
   def execPlanToProto(execPlan: ExecPlan): RemoteExecPlan = {
     val plan = execPlan.toExecPlanContainerProto()
@@ -71,7 +67,6 @@ object ProtoConverters extends StrictLogging {
       builder.addAllFlightPartitionsDenyList(qc.flightPartitionsDenyList.asJava)
       qc.plannerSelector.foreach(plannerSelector => builder.setPlannerSelector(plannerSelector))
       qc.recordContainerOverrides.foreach(overrides => builder.putRecordContainerOverrides(overrides._1, overrides._2))
-      builder.setSamplesScannedConfig(qc.samplesScannedConfig.toProto)
       builder.build()
     }
   }
@@ -99,154 +94,11 @@ object ProtoConverters extends StrictLogging {
         qc.getGrpcPartitionsDenyListList().asScala.toSet,
         qc.getFlightPartitionsDenyListList().asScala.toSet,
         if (qc.hasPlannerSelector) Option(qc.getPlannerSelector()) else None,
-        rcoIntMap,
-        samplesScannedConfig =
-          if (qc.hasSamplesScannedConfig) qc.getSamplesScannedConfig.fromProto
-          else SamplesScannedConfig()
+        rcoIntMap
       )
       queryConfig
     }
   }
-
-  implicit class SamplesScannedConfigToProtoConverter(config: SamplesScannedConfig) {
-    def toProto: GrpcMultiPartitionQueryService.SamplesScannedConfig = {
-      val builder = GrpcMultiPartitionQueryService.SamplesScannedConfig.newBuilder()
-      builder.setLeafSamplesEnabled(config.leafSamplesEnabled)
-      builder.setExecResultSamplesEnabled(config.execResultSamplesEnabled)
-      builder.setExecChildSamplesEnabled(config.execChildSamplesEnabled)
-      builder.setRvtSamplesEnabled(config.rvtSamplesEnabled)
-      builder.setRvtChildSamplesEnabled(config.rvtChildSamplesEnabled)
-      builder.setSrvSamplesEnabled(config.srvSamplesEnabled)
-      if (config.fixedRowMultiplier.isDefined) {
-        builder.setFixedRowMultiplier(config.fixedRowMultiplier.get)
-      }
-      builder.setDefaultRowMultiplier(config.defaultRowMultiplier)
-      builder.setHistogramRowMultiplier(config.histogramRowMultiplier)
-      builder.setExponentialHistogramRowMultiplier(config.exponentialHistogramRowMultiplier)
-      builder.setDefaultSamplesPerRow(config.defaultSamplesPerRow)
-      builder.setDefaultSamplesPerSeries(config.defaultSamplesPerSeries)
-      builder.setDefaultSamplesPerPartKeyByte(config.defaultSamplesPerPartKeyByte)
-      config.classToSamplesPerRow.foreach { case (clazz, mult) =>
-        builder.putClassToSamplesPerRow(clazz.getName, mult)
-      }
-      config.classToSamplesPerSeries.foreach { case (clazz, mult) =>
-        builder.putClassToSamplesPerSeries(clazz.getName, mult)
-      }
-      config.classToSamplesPerPartKeyByte.foreach { case (clazz, mult) =>
-        builder.putClassToSamplesPerPartKeyByte(clazz.getName, mult)
-      }
-      builder.setDefaultSamplesPerChildRow(config.defaultSamplesPerChildRow)
-      builder.setDefaultSamplesPerChildSeries(config.defaultSamplesPerChildSeries)
-      builder.setDefaultSamplesPerChildPartKeyByte(config.defaultSamplesPerChildPartKeyByte)
-      config.classToSamplesPerChildRow.foreach { case (clazz, mult) =>
-        builder.putClassToSamplesPerChildRow(clazz.getName, mult)
-      }
-      config.classToSamplesPerChildSeries.foreach { case (clazz, mult) =>
-        builder.putClassToSamplesPerChildSeries(clazz.getName, mult)
-      }
-      config.classToSamplesPerChildPartKeyByte.foreach { case (clazz, mult) =>
-        builder.putClassToSamplesPerChildPartKeyByte(clazz.getName, mult)
-      }
-      builder.build()
-    }
-  }
-
-  // scalastyle:off cyclomatic.complexity
-  implicit class SamplesScannedConfigFromProtoConverter(config: GrpcMultiPartitionQueryService.SamplesScannedConfig) {
-    def fromProto: SamplesScannedConfig = {
-      SamplesScannedConfig(
-        config.getLeafSamplesEnabled,
-        config.getExecResultSamplesEnabled,
-        config.getExecChildSamplesEnabled,
-        config.getRvtSamplesEnabled,
-        config.getRvtChildSamplesEnabled,
-        config.getSrvSamplesEnabled,
-        if (config.hasFixedRowMultiplier) Some(config.getFixedRowMultiplier) else None,
-        config.getDefaultRowMultiplier,
-        config.getHistogramRowMultiplier,
-        config.getExponentialHistogramRowMultiplier,
-        config.getDefaultSamplesPerRow,
-        config.getDefaultSamplesPerSeries,
-        config.getDefaultSamplesPerPartKeyByte,
-        config.getClassToSamplesPerRowMap.asScala
-          .filter { case (className, _) =>
-            val parseTry = Try(Class.forName(className))
-            if (!parseTry.isSuccess) {
-              unexpectedProtoClassNameCounter.increment()
-              logger.error("Could not find class while deserializing SamplesScannedConfig: " +
-                className, parseTry.failed.get)
-            }
-            parseTry.isSuccess
-          }
-          .map { case (className, mult) => Class.forName(className) -> mult.asInstanceOf[Double] }
-          .toMap,
-        config.getClassToSamplesPerSeriesMap.asScala
-          .filter { case (className, _) =>
-            val parseTry = Try(Class.forName(className))
-            if (!parseTry.isSuccess) {
-              unexpectedProtoClassNameCounter.increment()
-              logger.error("Could not find class while deserializing SamplesScannedConfig: " +
-                className, parseTry.failed.get)
-            }
-            parseTry.isSuccess
-          }
-          .map { case (className, mult) => Class.forName(className) -> mult.asInstanceOf[Double] }
-          .toMap,
-        config.getClassToSamplesPerPartKeyByteMap.asScala
-          .filter { case (className, _) =>
-            val parseTry = Try(Class.forName(className))
-            if (!parseTry.isSuccess) {
-              unexpectedProtoClassNameCounter.increment()
-              logger.error("Could not find class while deserializing SamplesScannedConfig: " +
-                className, parseTry.failed.get)
-            }
-            parseTry.isSuccess
-          }
-          .map { case (className, mult) => Class.forName(className) -> mult.asInstanceOf[Double] }
-          .toMap,
-        config.getDefaultSamplesPerChildRow,
-        config.getDefaultSamplesPerChildSeries,
-        config.getDefaultSamplesPerChildPartKeyByte,
-        config.getClassToSamplesPerChildRowMap.asScala
-          .filter { case (className, _) =>
-            val parseTry = Try(Class.forName(className))
-            if (!parseTry.isSuccess) {
-              unexpectedProtoClassNameCounter.increment()
-              logger.error("Could not find class while deserializing SamplesScannedConfig: " +
-                className, parseTry.failed.get)
-            }
-            parseTry.isSuccess
-          }
-          .map { case (className, mult) => Class.forName(className) -> mult.asInstanceOf[Double] }
-          .toMap,
-        config.getClassToSamplesPerChildSeriesMap.asScala
-          .filter { case (className, _) =>
-            val parseTry = Try(Class.forName(className))
-            if (!parseTry.isSuccess) {
-              unexpectedProtoClassNameCounter.increment()
-              logger.error("Could not find class while deserializing SamplesScannedConfig: " +
-                className, parseTry.failed.get)
-            }
-            parseTry.isSuccess
-          }
-          .map { case (className, mult) => Class.forName(className) -> mult.asInstanceOf[Double] }
-          .toMap,
-        config.getClassToSamplesPerChildPartKeyByteMap.asScala
-          .filter { case (className, _) =>
-            val parseTry = Try(Class.forName(className))
-            if (!parseTry.isSuccess) {
-              unexpectedProtoClassNameCounter.increment()
-              logger.error("Could not find class while deserializing SamplesScannedConfig: " +
-                className, parseTry.failed.get)
-            }
-            parseTry.isSuccess
-          }
-          .map { case (className, mult) => Class.forName(className) -> mult.asInstanceOf[Double] }
-          .toMap
-      )
-    }
-  }
-  // scalastyle:on cyclomatic.complexity
 
   // ChunkScanMethod
   implicit class ChunkScanMethodToProtoConverter(csm: filodb.core.store.ChunkScanMethod) {
