+++ grpc/src/main/protobuf/query_service.proto
@@ -50,27 +50,28 @@ enum FailoverMode {
 }
 
 message PlannerParams {
-    optional string applicationId                   = 1;
-    optional uint32 queryTimeoutMillis              = 2;
-    optional PerQueryLimits enforcedLimits          = 3;
-    optional PerQueryLimits warnLimits              = 4;
-    optional string queryOrigin                     = 5;
-    optional string queryOriginId                   = 6;
-    optional string queryPrincipal                  = 7;
-    optional bool timeSplitEnabled                  = 8;
-    optional uint64 minTimeRangeForSplitMs          = 9;
-    optional uint64 splitSizeMs                     = 10;
-    optional bool skipAggregatePresent              = 11;
-    optional bool processFailure                    = 12;
-    optional bool processMultiPartition             = 13;
-    optional bool allowPartialResults               = 14;
-    optional bool histogramMap                      = 15;
-    optional bool useProtoExecPlans                 = 16;
-    optional bool reduceShardKeyRegexFanout         = 17;
-    optional uint32 maxShardKeyRegexFanoutBatchSize = 18;
-    optional bool allowNestedAggregatePushdown      = 19;
-    repeated DownPartition downPartitions           = 20;
-    optional FailoverMode failoverMode              = 21;
+    optional string applicationId                       = 1;
+    optional uint32 queryTimeoutMillis                  = 2;
+    optional PerQueryLimits enforcedLimits              = 3;
+    optional PerQueryLimits warnLimits                  = 4;
+    optional string queryOrigin                         = 5;
+    optional string queryOriginId                       = 6;
+    optional string queryPrincipal                      = 7;
+    optional bool timeSplitEnabled                      = 8;
+    optional uint64 minTimeRangeForSplitMs              = 9;
+    optional uint64 splitSizeMs                         = 10;
+    optional bool skipAggregatePresent                  = 11;
+    optional bool processFailure                        = 12;
+    optional bool processMultiPartition                 = 13;
+    optional bool allowPartialResults                   = 14;
+    optional bool histogramMap                          = 15;
+    optional bool useProtoExecPlans                     = 16;
+    optional bool reduceShardKeyRegexFanout             = 17;
+    optional uint32 maxShardKeyRegexFanoutBatchSize     = 18;
+    optional bool allowNestedAggregatePushdown          = 19;
+    repeated DownPartition downPartitions               = 20;
+    optional FailoverMode failoverMode                  = 21;
+    optional SamplesScannedConfig samplesScannedConfig  = 22;
 }
 
 message Request {
@@ -455,8 +456,7 @@ message QueryConfig {
   repeated string grpcPartitionsDenyList     = 16;
   optional string plannerSelector            = 17;
   map<string,int32> recordContainerOverrides = 18;
-  SamplesScannedConfig samplesScannedConfig  = 19;
-  repeated string flightPartitionsDenyList   = 20;
+  repeated string flightPartitionsDenyList   = 19;
 }
 
 enum ChunkScanType {
